tested on pi and fixed bugs with energy/power displays
fixed button states which weren't working as they should - now there is some brief flashing of some buttons as their state is quickly toggled

+++ interactiveMode.py
@@ -95,8 +95,8 @@ class InteractiveMode():
         elif batteryLevelChange < 0:
             #discharging
             batteryLEDcolour = (int(self.LED_YELLOW_MAX[0] * (self.currentBatteryLevel/100)),
-            (self.LED_YELLOW_MAX[1] * (self.currentBatteryLevel/100)),
-            (self.LED_YELLOW_MAX[2] * (self.currentBatteryLevel/100)))
+            int((self.LED_YELLOW_MAX[1] * (self.currentBatteryLevel/100))),
+            int((self.LED_YELLOW_MAX[2] * (self.currentBatteryLevel/100))))
 
             self.board.setFuelCellLEDs(batteryLEDcolour)
             self.board.turnOffFuelCell()
